Update logic for trapping focus inside of popup.
`focusin` has been replaced with `keydown` to trap focus inside of popup.

Shift+tab on a first element inside of popup will bring focus to the last one, and tab on the last will bring focus to the first.

This change is also affecting tooltips so they will not have focus returned to the `open` element when clicked on a control in a background content outside of the popup.

This type implementation is closer to the W3C guidelines and pseudo code:
http://www.w3.org/TR/wai-aria-practices/#modal_dialog
http://www.w3.org/WAI/PF/aria-practices/#dialog_modal